+++ HISTORY.md
@@ -2,6 +2,7 @@
 
 ## Improvements
 * Refactor display of TileDB objects in Jupyter notebooks to be more readable [#1049](https://github.com/TileDB-Inc/TileDB-Py/pull/1049)
+* Improve documentation for `Filter`, `FilterList`, `VFS`, `FileIO`, and `Group` [#1043](https://github.com/TileDB-Inc/TileDB-Py/pull/1043)
 
 ## Bug Fixes
 * `Dim.shape` correctly errors out if type is not integer or datetime [#1055](https://github.com/TileDB-Inc/TileDB-Py/pull/1055)